+++ docs/changes.rst
@@ -17,6 +17,7 @@ Not yet released.
 * Added new generic check for placeholders.
 * Added support for WebExtension JSON placeholders.
 * Added support for flat XML format.
+* Exteded API with project, component and translation removal.
 
 weblate 3.8
 -----------